major refactor clean up
The change to using a dynamic mode generator angered the linter.  This
commit resolves the unknown global linter error by creating a static
global Modes to contain all the modes and their associated enter and
exit functions.
This commit also applies this to the utility functions, namespacing them
in a single global.  As part of this refactor, this commit also conforms
to lua style guidelines by using snake case on functions.